third_party/blink/web_tests/virtual/fetch-later/external/wpt/fetch/fetch-later/send-on-discard.tentative.https.window-expected.txt

This is a testharness.js-based test.
PASS A discarded document sends all its fetchLater requests.
FAIL A discarded document does not send an already aborted fetchLater request. assert_equals: Number of sent beacons does not match expected count: expected 1 but got 2
Harness: the test ran to completion.
